+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- next-header -->
 ## Unreleased
+## `0.5.0`
 
 - `LinearMap::new` and `LinearSet::new` are now `const`
 - Uses `foldhash` instead of `ahash` as default hasher.